requirements.txt
Django==2.2
djangorestframework==3.9.3
gunicorn==19.9.0
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
